Move functions to handler classes
