+++ common/util_functions.py
@@ -1,6 +1,7 @@
 import pyodbc
 import pandas as pd
 import os
+import time
 from python_ags4 import AGS4
 from PyQt5 import QtWidgets
 from PyQt5.QtWidgets import QMessageBox, QWidget
@@ -554,9 +555,37 @@ Check the AGS with "View data".''')
             print("No Lab or GEOL groups found - did this AGS contain CPT data? Check the data with 'View data'.")
             self.ags_table_reset()
 
-    def remove_match_id(self):
-        self.get_ags_tables()
+    def convert_excel(self):
+        try:
+            fname = QtWidgets.QFileDialog.getSaveFileName(self, "Save AGS as excel...", os.path.dirname(self.file_location), "Excel file *.xlsx;")
+        except:
+            fname = QtWidgets.QFileDialog.getSaveFileName(self, "Save AGS as excel...", os.getcwd(), "Excel file *.xlsx;")
+        
+        if fname[0] == '':
+            return
 
-        for table in self.ags_tables:
-            if "match_id" in self.tables[table]:
-                self.tables[table].drop(['match_id'], axis=1, inplace=True)
+        final_dataframes = [(k,v) for (k,v) in self.tables.items() if not v.empty]
+        final_dataframes = dict(final_dataframes)
+        empty_dataframes = [k for (k,v) in self.tables.items() if v.empty]
+
+        print(f"""------------------------------------------------------
+Saving AGS to excel file...
+------------------------------------------------------""")
+
+        #create the excel file with the first dataframe from dict, so pd.excelwriter can be called (can only be used on existing excel workbook to append more sheets)
+        if not len(final_dataframes.keys()) < 1:
+            next(iter(final_dataframes.values())).to_excel(f"{fname[0]}", sheet_name=(f"{next(iter(final_dataframes))}"), index=None, index_label=None)
+            final_writer = pd.ExcelWriter(f"{fname[0]}", engine="openpyxl", mode="a", if_sheet_exists="replace")
+        else:
+            print(f"All selected tables are empty! Please select others. Tables selected: {empty_dataframes}")
+            self.enable_buttons()
+            return
+
+        #for every key (table name) and value (table data) in the AGS, append to excel sheet and update progress bar, saving only at the end for performance
+        for (k,v) in final_dataframes.items():
+            print(f"Writing {k} to excel...")
+            v.to_excel(final_writer, sheet_name=(f"{str(k)}"), index=None, index_label=None)
+            time.sleep(0.01)
+        final_writer.close()
+
+        print(f"""AGS saved as Excel file: {fname[0]}""")
